Extract common class from response configs
As both class keep having more and more duplicated methods, it is time
to extract them to a parent class.

Only 'type' attribute is moved to the common class.
Others are left as is because moving them breaks serialization and many
tests. Doing something about it seems to be more effort than gain.
